Refactor .so bundling

+++ utils.c
@@ -55,9 +55,9 @@ bool bloom_check_or_add(bloom_t* bloom, uint8_t* hash, bool add) {
   b = mask - ((((uint32_t)(hash[3]) << 8) + hash[2]) & mask);
   c = mask - ((((uint32_t)(hash[5]) << 8) + hash[4]) & mask);
 
-  uint8_t ai = a / 8, av = 1 << (7 - (a % 8));
-  uint8_t bi = b / 8, bv = 1 << (7 - (b % 8));
-  uint8_t ci = c / 8, cv = 1 << (7 - (c % 8));
+  uint8_t ai = (uint8_t)(a / 8), av = (uint8_t)(1 << (7 - (a % 8)));
+  uint8_t bi = (uint8_t)(b / 8), bv = (uint8_t)(1 << (7 - (b % 8)));
+  uint8_t ci = (uint8_t)(c / 8), cv = (uint8_t)(1 << (7 - (c % 8)));
 
   if (add) {
     (*bloom)[ai] |= av;
